Fix two bugs: J00 doesn't work; Categories don't work
J00 didn't work because if a code doesn't have a dot, VistA adds
a dot to it. It turns out that you need to add a dot to a code that
doesn't contain dots (including categories) in order for the VistA
ICD-10 parsing code to work correctly.

Categories didn't work because of the dot issue; but there was
another problem. UPDATE^DIE did not create the "AFRAG" index in
757.033. A separate call to IX1^DIK was made to force the indexing.
Now categories work correctly.