Fix Y/N not working when DUZ("LANG")>1 when stuffing
Test case:
 >S DIE=666001,DA=1,DR=".02//Y",DUZ("LANG")=3 D ^DIE ; I'm speaking Spanish

YESNO: Y// ??
     Elija entre:
       1        Si
       2        No
YESNO: Y//

Fix in DIED.

DIALOGZ also had a bug in $$SETIN where it didn't translate the language
if either YES/NO is the same in the other language (e.g. in Spanish, NO
is the same as English, so we couldn't reproduce the bug).

New Fileman routine DIUCC required for new DIED.